backend/requirements.txt
# FastAPI - Modern, fast web framework for building APIs with automatic OpenAPI documentation
fastapi>=0.104.0

# Uvicorn - ASGI server for running FastAPI applications with high performance
uvicorn[standard]>=0.24.0

# Python-multipart - For handling file uploads and form data in FastAPI
python-multipart>=0.0.6

# PyPDF - For extracting text from PDF documents
pypdf>=3.17.0

# Python-docx - For extracting text from Microsoft Word documents
python-docx>=1.1.0

# TQDM - Progress bars for long-running operations (file processing, embeddings)
tqdm>=4.66.0

# Python-dotenv - For loading environment variables from .env files
python-dotenv>=1.0.0

# NumPy - Numerical computing library
numpy>=1.26.0

# SQLAlchemy - SQL toolkit and ORM for database operations
sqlalchemy>=2.0.0

# Requests - HTTP library for making API calls to Ollama
requests>=2.31.0

# Simple vector similarity search (alternative to FAISS)
scikit-learn>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.0.0
